requirements update .txt

streamlit
moviepy
gTTS
opencv-python
numpy
requests
pandas
transformers
sentencepiece
torch
Pillow
